requirements.txt
# Core Dependencies
tqdm
google-genai
pandas
pydantic
python-dotenv
datasets
huggingface-hub
pyarrow
openai

